requirements.txt
tensorflow
keras
opencv-contrib-python
numpy
flask
pillow